警告修正

+++ src/Engine/UnoEngine.cpp
@@ -556,8 +556,8 @@ void UnoEngine::InitializeImGui() {
 // === スムージングシステム実装 ===
 
 float UnoEngine::NormalizeAngle(float angle) {
-    while (angle > M_PI) angle -= 2.0f * M_PI;
-    while (angle < -M_PI) angle += 2.0f * M_PI;
+    while (angle > (float)M_PI) angle -= 2.0f * (float)M_PI;
+    while (angle < -(float)M_PI) angle += 2.0f * (float)M_PI;
     return angle;
 }
